fix: F69 3D viewer thread-safety hardening
- Add @Volatile to all Camera scalar fields (azimuth, elevation, distance,
  panX, panY, targetX, targetY, targetZ) — closes main↔GL visibility gap
- Add pendingCameraState to ModelRenderer consumed at top of onDrawFrame;
  route ModelViewerView.applyCameraState() and resetView() through it so
  the GL thread owns all camera mutations (eliminates data race on reset
  and state restore)
- Document @Volatile immutable-array contract on modelScale field
- Convert all Camera scalars to Double; downcast to Float only at shader
  uniform upload — eliminates float jitter at zoom extremes (matches
  SliceBeam approach)

+++ app/src/main/java/com/u1/slicer/ui/GcodeViewer3DScreen.kt
@@ -114,12 +114,12 @@ fun GcodeViewer3DScreen(
                     IconButton(onClick = {
                         val v = viewerView ?: return@IconButton
                         v.renderer.camera.apply {
-                            setTarget(135f, 135f, parsedGcode.layers.lastOrNull()?.z?.div(2) ?: 0f)
-                            distance = 400f
-                            elevation = 35f
-                            azimuth = -45f
-                            panX = 0f
-                            panY = 0f
+                            setTarget(135.0, 135.0, (parsedGcode.layers.lastOrNull()?.z?.div(2) ?: 0f).toDouble())
+                            distance = 400.0
+                            elevation = 35.0
+                            azimuth = -45.0
+                            panX = 0.0
+                            panY = 0.0
                         }
                         v.requestRender()
                     }) {
